LPS-85915 Move helper for FormInstanceRecord to service module
